fix: Do not reload actions when selecting purchase
